// src/app/interfaces/articulo.ts
export interface Articulo{
  
    cod_barras:string;
    cod_asociado: string;
    id_clasificacion: string;
    cod_interno: string;
    descripcion: string;
    descripcion_corta: string;
    cantidad_um: string;
    id_unidad: string;
    id_proveedor: string;
    precio_compra: string;
    utilidad: string;
    precio_venta: string;
    tipo_articulo: string;
    stock: string;
    stock_min: string;
    stock_max: string;
    iva: string;
    kit_fecha_ini: string;
    kit_fecha_fin: string;
    articulo_disponible: string;
    kit: string;
    fecha_registro: string;
    visible: string
    puntos: string;
    last_update_inventory: string;
    cve_producto: string;
}